refactor: Extracts formatter modules
