+++ lib/dialyxir/formatter/ignore_file.ex
@@ -1,0 +1,7 @@
+defmodule Dialyxir.Formatter.IgnoreFile do
+  @moduledoc false
+
+  def format({_tag, {file, _line}, {warning_name, _arguments}}) do
+    ~s({"#{file}", :#{warning_name}},)
+  end
+end
