+++ lib/dialyxir/formatter/raw.ex
@@ -1,0 +1,7 @@
+defmodule Dialyxir.Formatter.Raw do
+  @moduledoc false
+
+  def format(warning) do
+    inspect(warning, limit: :infinity)
+  end
+end
